Move power_balancer_energy scripts into energy_efficiency module
