+++ integration/experiment/Makefile.mk
@@ -39,6 +39,5 @@ EXTRA_DIST += integration/experiment/common_args.py \
 include integration/experiment/energy_efficiency/Makefile.mk
 include integration/experiment/frequency_sweep/Makefile.mk
 include integration/experiment/monitor/Makefile.mk
-include integration/experiment/power_balancer_energy/Makefile.mk
 include integration/experiment/power_sweep/Makefile.mk
 include integration/experiment/trace_analysis/Makefile.mk
