+++ integration/experiment/energy_efficiency/Makefile.mk
@@ -31,4 +31,7 @@
 
 EXTRA_DIST += integration/experiment/energy_efficiency/gen_plot_profile_comparison.py \
               integration/experiment/energy_efficiency/__init__.py \
+              integration/experiment/energy_efficiency/power_balancer_energy.py \
+              integration/experiment/energy_efficiency/run_power_balancer_energy_dgemm.py \
+              integration/experiment/energy_efficiency/run_power_balancer_energy_nekbone.py \
               # end
